Random number generator range change
This an update that includes a modified range of restaurants in an effort to mitigate program failure. 
The failure in testing resulted from a randomly generated number being plugged to the call which exceeded the actual amount of restaurants thrown back. Because of this, the program returned an "undefined" to the terminal during testing